+++ webapp/src/components/common/index.ts
@@ -1,4 +1,7 @@
 export { CommandPalette } from "./command-palette";
+export { CopyButton } from "./copy-button";
+export { DownloadButton } from "./download-button";
+export { FileUploadButton } from "./file-upload-button";
 export { HelpModal } from "./help-modal";
 export { SearchBar } from "./search-bar";
 export { SyntaxHighlighterComponent } from "./syntax-highlighter";