+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - feat: add syntax highlighter for data format tools
 - feat: add Progressive Web App (PWA) support
+- feat: add file upload and download support for data format tools
 
 ## [v1.3.0] - 2026-01-21